Rename SentVotesRow to SentStakesRow
