+++ src/modules/pos/validator/components/SentStakesRow/SentStakesRow.css
